LP-1873-1878-1881 Implement suggested changes after review
