+++ openedx/features/smart_referral/api_views.py
@@ -1,10 +1,15 @@
 from django.http import JsonResponse
 from edx_rest_framework_extensions.auth.session.authentication import SessionAuthenticationAllowInactiveUser
 from rest_framework import status
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+from rest_framework.status import HTTP_200_OK, HTTP_400_BAD_REQUEST
 from rest_framework.views import APIView
 
 from .helpers import sort_contacts_by_org_and_user_domain, get_platform_contacts_and_non_platform_contacts
+from .serializers import SmartReferralSerializer
+from .tasks import task_send_referral_and_toolkit_emails
 
 
 class FilterContactsAPIView(APIView):
@@ -28,3 +33,18 @@ class FilterContactsAPIView(APIView):
             'non_platform_contacts': non_platform_contacts,
         }
         return JsonResponse(response, status=status.HTTP_200_OK)
+
+
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+@authentication_classes([SessionAuthenticationAllowInactiveUser, ])
+def send_initial_emails_and_save_record(request):
+    serializer = SmartReferralSerializer(context={'request': request}, data=request.data, many=True, allow_empty=False)
+
+    if not serializer.is_valid():
+        return Response(serializer.errors, status=HTTP_400_BAD_REQUEST)
+
+    smart_referrals = serializer.save()
+    contact_emails = [referral.contact_email for referral in smart_referrals]
+    task_send_referral_and_toolkit_emails.delay(contact_emails=contact_emails, user_email=request.user.email)
+    return Response(status=HTTP_200_OK)
